+++ spec/dummy_app/app/models/tomato.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class Tomato < Plant
+  class << self
+    def sti_name
+      "tomato"
+    end
+  end
+end